Initial OTP plug-in implementation
This implements the main framework of a OTP authentication plug-in
for use by 389 DS / FreeIPA.  This plug-in follows the design mentioned
on the FreeIPA OTP design page:

    http://freeipa.org/page/V3/OTP

What remains is to fill in the actual TOTP authentication logic.  There
are also some comments for areas where we need to handle processing of critical
controls.  Additionally, there has been discussion about having the OTP plug-in
implement a control that can be used to return the type of authentication that
resulted in a successful bind.  This would allow the client to know if OTP or
password authentication was used when both methods are allowed.